Make main transactions page work
